added usage check for buffer tanks

+++ sophena/src/sophena/rcp/editors/basedata/buffers/BufferTankEditor.java
@@ -17,12 +17,15 @@ import org.eclipse.ui.forms.widgets.ScrolledForm;
 import org.eclipse.ui.forms.widgets.Section;
 
 import sophena.db.daos.RootEntityDao;
+import sophena.db.usage.SearchResult;
+import sophena.db.usage.UsageSearch;
 import sophena.model.BufferTank;
 import sophena.rcp.App;
 import sophena.rcp.Images;
 import sophena.rcp.M;
 import sophena.rcp.Numbers;
 import sophena.rcp.editors.Editor;
+import sophena.rcp.editors.basedata.UsageError;
 import sophena.rcp.utils.Actions;
 import sophena.rcp.utils.Editors;
 import sophena.rcp.utils.KeyEditorInput;
@@ -131,6 +134,11 @@ public class BufferTankEditor extends Editor {
 					"Soll der ausgewählte Pufferspeicher wirklich gelöscht werden?");
 			if (!doIt)
 				return;
+			List<SearchResult> usage = new UsageSearch(App.getDb()).of(b);
+			if (!usage.isEmpty()) {
+				UsageError.show(usage);
+				return;
+			}
 			try {
 				dao.delete(b);
 				buffers.remove(b);
